Reverse dependencies of modules, making types module prevailing over other main modules.
Add some concepts requiring compiler builtins. Curretly they are not guarded
with macros, which is the behaviour I'n not sure about. Managing the compiler
abilities should be considered more thoroughly. Also the types module was
simplified to make it more fundamental and independent of other modules. Tuple
module was put in the types extension module as it's implementation requires too
much stuff from other not so fundamental modules.